First try at rewriting for ZooKeeper configuration.
The idea is subclassing VoldemortConfig and setup the class with data from ZooKeeper.
The design is however questionable and should be reconsidered.

I am also not sure at current time how we should send the ZooKeeper startup string,
but it will probably be as command line parameter or some small config file.
This will probably have the format zk:hostname:port/path/to/chroot/to