Refuse to put keys that are metadata that can be written to zookeeper itself. This fixes a problem of eternal looping writes if setting metadata.
When using ZooKeeper, if a setData is called by a —set-metadata admin
command in voldemort, a put will be done in ZooKeeper. This triggers
the watch on all nodes, causing them to reload the file.

When doing rebalancing, set-metadata is called on every voldemort node,
resulting in n individual calls to setData for the ZooKeeper object.
every call to setData, triggers n watches, giving n*n watch triggers,
with possibility of chaining/looping.

Removed unused metadatastore reference.